server/requirements.txt
django-cors-headers==2.2.0
django==2.0.5
djangorestframework-camel-case==0.2.0
djangorestframework==3.8.2
drf-dynamic-fields==0.2.0
pyjwt==1.5.3
